refactor: profile modal
