+++ src/packages/@library/icons/SignOutIcon.svelte
